requirements.txt
Pillow==10.3.0
pyperclip==1.8.2
ttkbootstrap==1.10.1
